Add "legs" to player square
Two different leg sprites are intialized in approrpriate positions with
respect to the player square and then added as children to the player
square.

Doing this cleanly required rearranging the order of the lines that
control sprite positions and scene graph additions in the constructor. A
particular change is that the x and y coordinates for the bitmaps of
each sprite are no longer centered. This was originally done for the
player square so that it could be rotated about its center, and for
other objects so that they could be conveniently arranged about the
player square, but given that the player square is centered in the
display as of dc8f55fb3361015f34a19cf730ad58218c461e89, additional
centering seems redundant.